[EGO-19] Add "demo" Profile to control demo-server user creation
We want the ability to create users for a demo and ensure that their personal info is not saved to the demo DB (not visible in Users query in the data portal). Additionally, for a demo we want users to be admins by default so that they can access all features.

Creation of "demo" profile in application.yml, by running the server with this profile enabled the application will control if the user is saved and what their role is on creation. If profile is not enabled, server will perform user creation as normal.

https://github.com/overture-stack/ego/issues/19