[Ego-31] CircleCI Deploy Script and Dependencies First pass at deploy script, requires testing.
Also added was a env_template.sh file. The intention is to record all environment variables required on the server in a convenient script that can be modified to a local version with the environment specific values. The .gitignore file is updated to ignore any env.sh file with specific version variables.

The start-server.sh and stop-server.sh scripts depend on the environment variables listed in the env_template.sh file.

Specific instructions for other steps to take before the CircleCI deploy will work are listed at the start of the deploy.sh script.