Refactor User Event and Receiver Addresses two problems: 1. Event name wasn't unique, was reporting as "UPDATE". By adding the class name to the event identifier there won't be confusion between different update events. 2. The EventBus would break on dispatch of an event if the notify event had a different Object type than one of its subscribers. Some testing showed that this would prevent events with the matching data type from receiving the event - stopped dispatch completely. Careful coding would ensure all events had a data type to match the consumers, but this but the safer way to handle it is to make all Consumers of type Event<?> so that they will accept all Event payloads, and then to make the consumer explicitly cast the payload to the type they expect.
Now that the consumers are safe to receive all Events, we want to make it straightforward for users of our Events to pass the expected data type. To do this, we moved the eventBus.notify code into a method in the UserEvent class. Now the invoker of the event doesn't need to know abotu eventBus or any reactor code - only needs to call UserEvents.update(user). This also ensures that the correct type of object is passed in the Event payload.