Refactoring:    1) Renamed params class "Scope" to "ScopeName" (it contains the names of a Policy, and a policy mask).    2) Created a new class called "Scope" to hold a Policy object, and a PolicyMask object.    3) Renamed the DTO class "TokenScope" to "TokenScopeResponse".    4) Created a new class called "TokenScope" to map the "tokenscope" table.
Enhancement: In issue_token, we now save the PolicyMask in addition to the Policy in the tokenscope table, so that we can issue tokens with READ access for a given Policy, even if the user has been granted WRITE access.

Enhancement: In check_token, we now check the access levels of the token against the current access level of the user.

Enhancement: In class User, added two functions: allowedScopes() (which returns from a given set of Scope objects the ones that the user has access to), and missingScopes(), the opposite function.

Enhancement: In class PolicyMask, added a static boolean function called "allows()", which for a given pair of PolicyMasks, returns true if the first mask allows access to the second. (eg. WRITE allows READ, but READ does NOT allow WRITE)

Enhancement: In ScopeAwareOauth2Request, we now validate against the set of Scopes, not policies.

Enhancement: In TokenService, added functions getScopes(), and getScope(),
which take a string consisting of a policyname, and a policymask name, and
returns a Scope object containing the appropriate Policy and PolicyMask objects,by looking up the right Policy from the PolicyService.

Enhancement: Also in TokenService, add a missingScopes() method, which takes
a User object, and a set of scope names, and returns those scopes, if any, which
that user does not have access to.

TODO:
1) Write unit tests to ensure the access levels are correctly computed.
2) Find a better name than ScopeName for the PolicyId String + PolicyMask String ...
3) Integration test code to make sure it still runs...